+++ charts/camunda-platform/test/unit/console/goldenfiles_test.go
@@ -17,10 +17,8 @@ package console
 import (
 	"camunda-platform-helm/charts/camunda-platform/test/unit/utils"
 	"path/filepath"
-	"strings"
 	"testing"
 
-	"github.com/gruntwork-io/terratest/modules/random"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 )
@@ -41,7 +39,7 @@ func TestGoldenDefaultsTemplate(t *testing.T) {
 		suite.Run(t, &utils.TemplateGoldenTest{
 			ChartPath:      chartPath,
 			Release:        "camunda-platform-test",
-			Namespace:      "camunda-platform-" + strings.ToLower(random.UniqueId()),
+			Namespace:      "camunda-platform",
 			GoldenFileName: name,
 			Templates:      []string{"templates/console/" + name + ".yaml"},
 			SetValues: map[string]string{