+++ charts/camunda-platform/RELEASE-NOTES.md
@@ -2,6 +2,10 @@ The changelog is automatically generated using [git-chglog](https://github.com/g
 and it follows [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/) format.
 
 
-<a name="camunda-platform-9.1.2"></a>
-## [camunda-platform-9.1.2](https://github.com/camunda/camunda-platform-helm/compare/camunda-platform-9.1.1...camunda-platform-9.1.2) (2024-02-15)
+<a name="camunda-platform-9.3.1"></a>
+## [camunda-platform-9.3.1](https://github.com/camunda/camunda-platform-helm/compare/camunda-platform-9.3.0...camunda-platform-9.3.1) (2024-03-15)
 
+### Fix
+
+* use the correct console.image.tag instead of SNAPSHOT
+* set identity automountServiceAccountToken to true instead of True